chore: 디렉토리 분류
