+++ src/main/java/domain/piece/AbstractPiece.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.piece;
 
+import domain.Score;
+import domain.Team;
 import java.util.Objects;
 
 /**
