+++ src/main/java/domain/piece/Chariot.java
@@ -1,23 +1,24 @@
-package domain;
+package domain.piece;
 
+import domain.Score;
+import domain.Team;
+import domain.position.Distance;
+import domain.position.Point;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Cannon extends AbstractPiece {
+public class Chariot extends AbstractPiece {
 
-    public Cannon(final Team team, final Score score) {
+    public Chariot(final Team team, final Score score) {
         super(team, score);
     }
 
-    /*
-        포는 말 1개를 건너 뛰어야만 이동 가능하므로 target은 prev에서 한 칸 옮긴 위치에서 시작한다.
-         */
     @Override
     public List<Point> getPossiblePoint(Point prev, Point newPoint) {
         List<Point> possiblePoint = new ArrayList<>();
 
-        final int x = prev.calculateSubtractionX(newPoint);
-        final int y = prev.calculateSubtractionY(newPoint);
+        int x = prev.calculateSubtractionX(newPoint);
+        int y = prev.calculateSubtractionY(newPoint);
         if (x > 0) {
             Point target = prev;
             for (int i = 0; i < prev.distanceToMaxX(); i++) {
@@ -63,11 +64,9 @@ public class Cannon extends AbstractPiece {
 
     @Override
     public boolean isMovable(final Distance distance) {
-        final int absoluteX = Math.abs(distance.x());
-        final int absoluteY = Math.abs(distance.y());
-        if (absoluteX >= 2 && absoluteY == 0) {
+        if (distance.x() == 0 && distance.y() != 0) {
             return true;
         }
-        return absoluteX == 0 && absoluteY >= 2;
+        return distance.x() != 0 && distance.y() == 0;
     }
 }
