+++ src/main/java/domain/piece/Elephant.java
@@ -1,11 +1,15 @@
-package domain;
+package domain.piece;
 
+import domain.Score;
+import domain.Team;
+import domain.position.Distance;
+import domain.position.Point;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Horse extends AbstractPiece {
+public class Elephant extends AbstractPiece {
 
-    public Horse(final Team team, final Score score) {
+    public Elephant(final Team team, final Score score) {
         super(team, score);
     }
 
@@ -16,36 +20,46 @@ public class Horse extends AbstractPiece {
         int x = prev.calculateSubtractionX(newPoint);
         int y = prev.calculateSubtractionY(newPoint);
         if (x > 0 && y > 0) {
-            final Point point = prev.up().rightUp();
+            final Point point = prev.up().rightUp().rightUp();
 
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.up());
+                possiblePoint.add(prev.up().rightUp());
             } else {
                 possiblePoint.add(prev.right());
+                possiblePoint.add(prev.right().rightUp());
             }
         }
         if (x < 0 && y > 0) {
-            final Point point = prev.up().leftUp();
+            final Point point = prev.up().leftUp().leftUp();
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.up());
+                possiblePoint.add(prev.up().leftUp());
             } else {
                 possiblePoint.add(prev.left());
+                possiblePoint.add(prev.left().leftUp());
             }
+
         }
         if (x < 0 && y < 0) {
-            final Point point = prev.left().leftDown();
+            final Point point = prev.left().leftDown().leftDown();
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.left());
+                possiblePoint.add(prev.left().leftDown());
             } else {
                 possiblePoint.add(prev.down());
+                possiblePoint.add(prev.down().leftDown());
             }
+
         }
         if (x > 0 && y < 0) {
-            final Point point = prev.down().rightDown();
+            final Point point = prev.down().rightDown().rightDown();
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.down());
+                possiblePoint.add(prev.down().rightDown());
             } else {
                 possiblePoint.add(prev.right());
+                possiblePoint.add(prev.right().rightDown());
             }
         }
         return possiblePoint;
@@ -55,9 +69,9 @@ public class Horse extends AbstractPiece {
     public boolean isMovable(final Distance distance) {
         int absoluteX = Math.abs(distance.x());
         int absoluteY = Math.abs(distance.y());
-        if (absoluteX == 1 && absoluteY == 2) {
+        if (absoluteX == 2 && absoluteY == 3) {
             return true;
         }
-        return absoluteX == 2 && absoluteY == 1;
+        return absoluteX == 3 && absoluteY == 2;
     }
 }
