+++ src/main/java/domain/piece/Horse.java
@@ -1,11 +1,15 @@
-package domain;
+package domain.piece;
 
+import domain.Score;
+import domain.Team;
+import domain.position.Distance;
+import domain.position.Point;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Elephant extends AbstractPiece {
+public class Horse extends AbstractPiece {
 
-    public Elephant(final Team team, final Score score) {
+    public Horse(final Team team, final Score score) {
         super(team, score);
     }
 
@@ -16,46 +20,36 @@ public class Elephant extends AbstractPiece {
         int x = prev.calculateSubtractionX(newPoint);
         int y = prev.calculateSubtractionY(newPoint);
         if (x > 0 && y > 0) {
-            final Point point = prev.up().rightUp().rightUp();
+            final Point point = prev.up().rightUp();
 
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.up());
-                possiblePoint.add(prev.up().rightUp());
             } else {
                 possiblePoint.add(prev.right());
-                possiblePoint.add(prev.right().rightUp());
             }
         }
         if (x < 0 && y > 0) {
-            final Point point = prev.up().leftUp().leftUp();
+            final Point point = prev.up().leftUp();
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.up());
-                possiblePoint.add(prev.up().leftUp());
             } else {
                 possiblePoint.add(prev.left());
-                possiblePoint.add(prev.left().leftUp());
             }
-
         }
         if (x < 0 && y < 0) {
-            final Point point = prev.left().leftDown().leftDown();
+            final Point point = prev.left().leftDown();
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.left());
-                possiblePoint.add(prev.left().leftDown());
             } else {
                 possiblePoint.add(prev.down());
-                possiblePoint.add(prev.down().leftDown());
             }
-
         }
         if (x > 0 && y < 0) {
-            final Point point = prev.down().rightDown().rightDown();
+            final Point point = prev.down().rightDown();
             if (point.equals(newPoint)) {
                 possiblePoint.add(prev.down());
-                possiblePoint.add(prev.down().rightDown());
             } else {
                 possiblePoint.add(prev.right());
-                possiblePoint.add(prev.right().rightDown());
             }
         }
         return possiblePoint;
@@ -65,9 +59,9 @@ public class Elephant extends AbstractPiece {
     public boolean isMovable(final Distance distance) {
         int absoluteX = Math.abs(distance.x());
         int absoluteY = Math.abs(distance.y());
-        if (absoluteX == 2 && absoluteY == 3) {
+        if (absoluteX == 1 && absoluteY == 2) {
             return true;
         }
-        return absoluteX == 3 && absoluteY == 2;
+        return absoluteX == 2 && absoluteY == 1;
     }
 }
