+++ src/main/java/domain/piece/Piece.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.piece;
 
+import domain.Score;
+import domain.position.Distance;
+import domain.position.Point;
 import java.util.List;
 
 public interface Piece {
