+++ src/main/java/domain/position/Distance.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.position;
 
 public record Distance(int x, int y) {
 }
