+++ src/test/java/domain/AbstractPieceTest.java
@@ -1,5 +1,9 @@
 package domain;
 
+import domain.piece.AbstractPiece;
+import domain.piece.PieceFactory;
+import domain.position.Distance;
+import domain.position.Point;
 import java.util.List;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
