+++ src/test/java/domain/BoardTest.java
@@ -4,6 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.piece.Chariot;
+import domain.piece.PieceFactory;
+import domain.position.Point;
+import domain.position.Position;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
