+++ src/test/java/domain/ChariotTest.java
@@ -2,6 +2,10 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.piece.Chariot;
+import domain.piece.PieceFactory;
+import domain.position.Distance;
+import domain.position.Point;
 import java.util.List;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
